Factor out prop types
